refactor(backend): extract sync pipeline to utils
Move decode→VAD→STT→merge pipeline from routers/sync.py into
utils/sync/pipeline.py so the router stays thin and utils never
imports routers. Update workflow contract sources and test anchors.

Verification: pytest tests/unit/test_sync_v2.py
tests/unit/test_sync_cloud_tasks.py tests/unit/test_sync_ordered_assignment.py
tests/unit/test_sync_silent_failure.py tests/unit/test_workflow_contracts.py
tests/unit/test_utils_import_boundary.py (248 passed)
