+++ backend/tests/unit/test_sync_silent_failure.py
@@ -30,13 +30,13 @@ class TestProcessSegmentErrorHandling:
     """Verify process_segment has proper error handling after the fix."""
 
     @staticmethod
-    def _read_sync_source():
-        sync_path = os.path.join(os.path.dirname(__file__), '..', '..', 'routers', 'sync.py')
-        return _read_text(sync_path)
+    def _read_pipeline_source():
+        pipeline_path = os.path.join(os.path.dirname(__file__), '..', '..', 'utils', 'sync', 'pipeline.py')
+        return _read_text(pipeline_path)
 
     def test_process_segment_has_try_except(self):
         """process_segment() must wrap its body in try/except to catch all errors."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         start = source.index('def process_segment(')
         next_def = source.index('\ndef ', start + 1)
         func_body = source[start:next_def]
@@ -46,7 +46,7 @@ class TestProcessSegmentErrorHandling:
 
     def test_process_segment_treats_empty_words_as_success(self):
         """When Deepgram returns no words, treat it as a successful empty segment."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         start = source.index('def process_segment(')
         next_def = source.index('\ndef ', start + 1)
         func_body = source[start:next_def]
@@ -58,7 +58,7 @@ class TestProcessSegmentErrorHandling:
 
     def test_process_segment_warns_on_empty_postprocessed(self):
         """When words exist but postprocessing yields nothing, log warning (not error)."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         start = source.index('def process_segment(')
         next_def = source.index('\ndef ', start + 1)
         func_body = source[start:next_def]
@@ -67,7 +67,7 @@ class TestProcessSegmentErrorHandling:
 
     def test_process_segment_collects_errors_on_exception(self):
         """When an exception occurs, it must be caught and appended to errors."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         start = source.index('def process_segment(')
         next_def = source.index('\ndef ', start + 1)
         func_body = source[start:next_def]
@@ -76,7 +76,7 @@ class TestProcessSegmentErrorHandling:
 
     def test_process_segment_uses_lock_for_thread_safety(self):
         """Shared state mutations must be protected by a lock."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         start = source.index('def process_segment(')
         next_def = source.index('\ndef ', start + 1)
         func_body = source[start:next_def]
@@ -88,7 +88,7 @@ class TestProcessSegmentErrorHandling:
 
     def test_process_segment_accepts_lock_and_errors_params(self):
         """process_segment must accept lock and errors as parameters."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         start = source.index('def process_segment(')
         sig_end = source.index('):', start) + 2
         signature = source[start:sig_end]
@@ -514,13 +514,13 @@ class TestSegmentDeduplication:
     """Verifies that retried segments are deduplicated in the merge path."""
 
     @staticmethod
-    def _read_sync_source():
-        sync_path = os.path.join(os.path.dirname(__file__), '..', '..', 'routers', 'sync.py')
-        return _read_text(sync_path)
+    def _read_pipeline_source():
+        pipeline_path = os.path.join(os.path.dirname(__file__), '..', '..', 'utils', 'sync', 'pipeline.py')
+        return _read_text(pipeline_path)
 
     def test_merge_path_has_dedup_logic(self):
         """process_segment merge path must deduplicate before appending."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         start = source.index('def process_segment(')
         next_def = source.index('\ndef ', start + 1)
         func_body = source[start:next_def]
@@ -533,7 +533,7 @@ class TestSegmentDeduplication:
 
     def test_dedup_uses_timestamp_rounding(self):
         """Dedup must round timestamps to avoid float precision issues."""
-        source = self._read_sync_source()
+        source = self._read_pipeline_source()
         assert 'round(' in source, "Must round timestamps for reliable comparison"
 
     def test_dedup_logic_correctness(self):
@@ -640,8 +640,9 @@ class TestProcessSegmentReal:
     def setup_class(cls):
         # Save originals
         cls._saved_modules = {name: sys.modules.get(name) for name in _STUB_MODULES}
-        # Also save routers.sync if already imported
-        cls._saved_modules['routers.sync'] = sys.modules.get('routers.sync')
+        # Also save pipeline if already imported
+        cls._saved_modules['utils.sync.pipeline'] = sys.modules.get('utils.sync.pipeline')
+        cls._saved_modules['utils.sync'] = sys.modules.get('utils.sync')
 
         # Install stubs
         for mod_name in _STUB_MODULES:
@@ -747,18 +748,19 @@ class TestProcessSegmentReal:
         sys.modules['models.sync_audio'].AudioPrecacheResponse = _AudioPrecacheResponse
         sys.modules['models.sync_audio'].AudioUrlsResponse = _AudioUrlsResponse
 
-        # Import under stubs
-        from routers.sync import process_segment
-        import routers.sync as sync_mod
+        sync_pkg = ModuleType('utils.sync')
+        sync_pkg.__path__ = [os.path.join(os.path.dirname(__file__), '..', '..', 'utils', 'sync')]
+        sys.modules['utils.sync'] = sync_pkg
+        sys.modules.pop('utils.sync.pipeline', None)
 
-        sync_mod.submit_with_context = MagicMock()
+        # Import under stubs
+        from utils.sync.pipeline import process_segment
 
         cls._process_segment = staticmethod(process_segment)
 
     @classmethod
     def teardown_class(cls):
-        # Remove routers.sync so it can be re-imported cleanly
-        sys.modules.pop('routers.sync', None)
+        sys.modules.pop('utils.sync.pipeline', None)
         # Restore original modules
         for name, orig in cls._saved_modules.items():
             if orig is None:
@@ -778,10 +780,10 @@ class TestProcessSegmentReal:
         errors = []
         lock = threading.Lock()
 
-        with patch('routers.sync.prerecorded', return_value=([], 'en')), patch(
-            'routers.sync.delete_syncing_temporal_file'
-        ), patch('routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'), patch(
-            'routers.sync.time.sleep'
+        with patch('utils.sync.pipeline.prerecorded', return_value=([], 'en')), patch(
+            'utils.sync.pipeline.delete_syncing_temporal_file'
+        ), patch('utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'), patch(
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
 
@@ -803,12 +805,12 @@ class TestProcessSegmentReal:
         errors = []
         lock = threading.Lock()
 
-        with patch('routers.sync.prerecorded', return_value=([{'text': 'um'}], 'en')), patch(
-            'routers.sync.postprocess_words', return_value=[]
-        ), patch('routers.sync.delete_syncing_temporal_file'), patch(
-            'routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'
+        with patch('utils.sync.pipeline.prerecorded', return_value=([{'text': 'um'}], 'en')), patch(
+            'utils.sync.pipeline.postprocess_words', return_value=[]
+        ), patch('utils.sync.pipeline.delete_syncing_temporal_file'), patch(
+            'utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'
         ), patch(
-            'routers.sync.time.sleep'
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
 
@@ -826,10 +828,10 @@ class TestProcessSegmentReal:
         errors = []
         lock = threading.Lock()
 
-        with patch('routers.sync.prerecorded', side_effect=ConnectionError('timeout')), patch(
-            'routers.sync.delete_syncing_temporal_file'
-        ), patch('routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'), patch(
-            'routers.sync.time.sleep'
+        with patch('utils.sync.pipeline.prerecorded', side_effect=ConnectionError('timeout')), patch(
+            'utils.sync.pipeline.delete_syncing_temporal_file'
+        ), patch('utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'), patch(
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
 
@@ -857,18 +859,18 @@ class TestProcessSegmentReal:
         mock_conv = MagicMock()
         mock_conv.id = 'conv-abc123'
 
-        with patch('routers.sync.prerecorded', return_value=([{'text': 'hello'}], 'en')), patch(
-            'routers.sync.postprocess_words', return_value=[real_segment]
-        ), patch('routers.sync.get_timestamp_from_path', return_value=1700000000.0), patch(
-            'routers.sync.get_closest_conversation_to_timestamps', return_value=None
+        with patch('utils.sync.pipeline.prerecorded', return_value=([{'text': 'hello'}], 'en')), patch(
+            'utils.sync.pipeline.postprocess_words', return_value=[real_segment]
+        ), patch('utils.sync.pipeline.get_timestamp_from_path', return_value=1700000000.0), patch(
+            'utils.sync.pipeline.get_closest_conversation_to_timestamps', return_value=None
         ), patch(
-            'routers.sync.process_conversation', return_value=mock_conv
+            'utils.sync.pipeline.process_conversation', return_value=mock_conv
         ), patch(
-            'routers.sync.delete_syncing_temporal_file'
+            'utils.sync.pipeline.delete_syncing_temporal_file'
         ), patch(
-            'routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'
+            'utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'
         ), patch(
-            'routers.sync.time.sleep'
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
 
@@ -900,18 +902,18 @@ class TestProcessSegmentReal:
         mock_conv = MagicMock()
         mock_conv.id = 'conv-success'
 
-        with patch('routers.sync.prerecorded', side_effect=mock_deepgram_mixed), patch(
-            'routers.sync.postprocess_words', return_value=[real_segment]
-        ), patch('routers.sync.get_timestamp_from_path', return_value=1700000000.0), patch(
-            'routers.sync.get_closest_conversation_to_timestamps', return_value=None
+        with patch('utils.sync.pipeline.prerecorded', side_effect=mock_deepgram_mixed), patch(
+            'utils.sync.pipeline.postprocess_words', return_value=[real_segment]
+        ), patch('utils.sync.pipeline.get_timestamp_from_path', return_value=1700000000.0), patch(
+            'utils.sync.pipeline.get_closest_conversation_to_timestamps', return_value=None
         ), patch(
-            'routers.sync.process_conversation', return_value=mock_conv
+            'utils.sync.pipeline.process_conversation', return_value=mock_conv
         ), patch(
-            'routers.sync.delete_syncing_temporal_file'
+            'utils.sync.pipeline.delete_syncing_temporal_file'
         ), patch(
-            'routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'
+            'utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'
         ), patch(
-            'routers.sync.time.sleep'
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
 
@@ -959,18 +961,18 @@ class TestProcessSegmentReal:
 
         existing_conv['finished_at'] = datetime.fromtimestamp(1700000005.0, tz=timezone.utc)
 
-        with patch('routers.sync.prerecorded', return_value=([{'text': 'hello'}], 'en')), patch(
-            'routers.sync.postprocess_words', return_value=[mock_segment]
-        ), patch('routers.sync.get_timestamp_from_path', return_value=1700000000.0), patch(
-            'routers.sync.get_closest_conversation_to_timestamps', return_value=existing_conv
+        with patch('utils.sync.pipeline.prerecorded', return_value=([{'text': 'hello'}], 'en')), patch(
+            'utils.sync.pipeline.postprocess_words', return_value=[mock_segment]
+        ), patch('utils.sync.pipeline.get_timestamp_from_path', return_value=1700000000.0), patch(
+            'utils.sync.pipeline.get_closest_conversation_to_timestamps', return_value=existing_conv
         ), patch(
-            'routers.sync.update_conversation_segments'
+            'utils.sync.pipeline.update_conversation_segments'
         ) as mock_update, patch(
-            'routers.sync.delete_syncing_temporal_file'
+            'utils.sync.pipeline.delete_syncing_temporal_file'
         ), patch(
-            'routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'
+            'utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'
         ), patch(
-            'routers.sync.time.sleep'
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
 
@@ -994,10 +996,10 @@ class TestProcessSegmentReal:
         lock = threading.Lock()
 
         # Run 3 segments that all return empty words (silence)
-        with patch('routers.sync.prerecorded', return_value=([], 'en')), patch(
-            'routers.sync.delete_syncing_temporal_file'
-        ), patch('routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'), patch(
-            'routers.sync.time.sleep'
+        with patch('utils.sync.pipeline.prerecorded', return_value=([], 'en')), patch(
+            'utils.sync.pipeline.delete_syncing_temporal_file'
+        ), patch('utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'), patch(
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
 
@@ -1031,12 +1033,12 @@ class TestProcessSegmentReal:
         lock = threading.Lock()
 
         with patch(
-            'routers.sync.prerecorded',
+            'utils.sync.pipeline.prerecorded',
             side_effect=RuntimeError('Deepgram transcription failed after 2 attempts: timeout'),
-        ), patch('routers.sync.delete_syncing_temporal_file'), patch(
-            'routers.sync.get_syncing_file_temporal_signed_url', return_value='https://fake'
+        ), patch('utils.sync.pipeline.delete_syncing_temporal_file'), patch(
+            'utils.sync.pipeline.get_syncing_file_temporal_signed_url', return_value='https://fake'
         ), patch(
-            'routers.sync.time.sleep'
+            'utils.sync.pipeline.time.sleep'
         ):
             from models.conversation_enums import ConversationSource
