+++ backend/tests/unit/test_sync_v2.py
@@ -20,6 +20,15 @@ from io import BytesIO
 from unittest.mock import MagicMock, patch
 
 import pytest
+
+PIPELINE_PATH = os.path.join(os.path.dirname(__file__), '..', '..', 'utils', 'sync', 'pipeline.py')
+
+
+def _read_pipeline_source():
+    with open(PIPELINE_PATH, encoding='utf-8') as f:
+        return f.read()
+
+
 from pydantic import BaseModel
 
 # ---------------------------------------------------------------------------
@@ -110,9 +119,9 @@ class TestSyncV2Structure:
 
     def test_v2_uses_job_specific_directory(self):
         """v2 must use syncing/{uid}/{job_id}/ to avoid concurrency conflicts."""
-        source = self._read_sync_source()
-        assert '_retrieve_file_paths_v2' in source
-        start = source.index('def _retrieve_file_paths_v2')
+        source = _read_pipeline_source()
+        assert '_retrieve_file_paths_v2' in _read_pipeline_source()
+        start = _read_pipeline_source().index('def _retrieve_file_paths_v2')
         next_def = source.index('\ndef ', start + 1)
         func_body = source[start:next_def]
 
@@ -122,7 +131,7 @@ class TestSyncV2Structure:
 
     def test_v2_background_has_cleanup(self):
         """Background worker must clean up files in finally block."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
         next_boundary = source.find('\n@router.', start + 1)
         if next_boundary == -1:
@@ -135,7 +144,7 @@ class TestSyncV2Structure:
 
     def test_v2_background_records_dg_after_processing(self):
         """DG usage must be recorded after processing, not before."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
         next_boundary = source.find('\n@router.', start + 1)
         if next_boundary == -1:
@@ -148,7 +157,7 @@ class TestSyncV2Structure:
 
     def test_v2_background_does_decode_and_vad(self):
         """Background worker must run decode and VAD (#7281 — moved from inline)."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
         next_boundary = source.find('\n@router.', start + 1)
         if next_boundary == -1:
@@ -162,7 +171,7 @@ class TestSyncV2Structure:
 
     def test_v2_background_has_stage_heartbeats(self):
         """Background worker must heartbeat with stage info during decode and VAD."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
         next_boundary = source.find('\n@router.', start + 1)
         if next_boundary == -1:
@@ -194,7 +203,7 @@ class TestSyncV2Structure:
 
     def test_v2_bg_worker_fetches_prefs_and_cache(self):
         """Background worker must fetch transcription prefs and build person embeddings cache."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
         next_boundary = source.find('\n@router.', start + 1)
         if next_boundary == -1:
@@ -206,9 +215,11 @@ class TestSyncV2Structure:
 
     def test_v2_bg_worker_forwards_private_cloud_sync_enabled(self):
         """Background worker must forward private cloud sync intent into process_segment."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
-        next_boundary = source.find('\n@router.', start + 1)
+        next_boundary = source.find('\nasync def ', start + 1)
+        if next_boundary == -1:
+            next_boundary = source.find('\ndef ', start + 1)
         if next_boundary == -1:
             next_boundary = len(source)
         func_body = source[start:next_boundary]
@@ -442,11 +453,11 @@ class TestFullPipelineBackground:
 
     @staticmethod
     def _get_bg_func_body():
-        sync_path = os.path.join(os.path.dirname(__file__), '..', '..', 'routers', 'sync.py')
-        with open(sync_path, encoding='utf-8') as f:
-            source = f.read()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
-        next_boundary = source.find('\n@router.', start + 1)
+        next_boundary = source.find('\nasync def ', start + 1)
+        if next_boundary == -1:
+            next_boundary = source.find('\ndef ', start + 1)
         if next_boundary == -1:
             next_boundary = len(source)
         return source[start:next_boundary]
@@ -802,11 +813,11 @@ class TestAsyncCoordinatorStructure:
 
     @staticmethod
     def _get_bg_func_body():
-        sync_path = os.path.join(os.path.dirname(__file__), '..', '..', 'routers', 'sync.py')
-        with open(sync_path, encoding='utf-8') as f:
-            source = f.read()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
-        next_boundary = source.find('\n@router.', start + 1)
+        next_boundary = source.find('\nasync def ', start + 1)
+        if next_boundary == -1:
+            next_boundary = source.find('\ndef ', start + 1)
         if next_boundary == -1:
             next_boundary = len(source)
         return source[start:next_boundary]
@@ -930,13 +941,13 @@ class TestAsyncCoordinatorSemaphore:
 
     def test_semaphore_delegates_to_http_client(self):
         """Semaphore must use http_client._get_semaphore (CLAUDE.md rule 4)."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         assert '_get_sync_pipeline_semaphore()' in source, "Must use loop-scoped semaphore getter"
         assert '_get_semaphore' in source, "Must delegate to http_client._get_semaphore"
 
     def test_semaphore_limit_is_16(self):
         """Semaphore cap must be 16 (2x the old 8-slot postprocess_executor)."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('def _get_sync_pipeline_semaphore')
         end = source.find('\ndef ', start + 1)
         if end == -1:
@@ -962,11 +973,11 @@ class TestAsyncCoordinatorScenarios:
 
     @staticmethod
     def _get_bg_func_body():
-        sync_path = os.path.join(os.path.dirname(__file__), '..', '..', 'routers', 'sync.py')
-        with open(sync_path, encoding='utf-8') as f:
-            source = f.read()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
-        next_boundary = source.find('\n@router.', start + 1)
+        next_boundary = source.find('\nasync def ', start + 1)
+        if next_boundary == -1:
+            next_boundary = source.find('\ndef ', start + 1)
         if next_boundary == -1:
             next_boundary = len(source)
         return source[start:next_boundary]
@@ -1320,7 +1331,6 @@ class TestAsyncCoordinatorBehavioral:
             'utils.log_sanitizer',
             'utils.http_client',
             'utils.request_validation',
-            'utils.sync',
             'utils.sync.files',
             'utils.sync.playback',
             'utils.speaker_assignment',
@@ -1336,6 +1346,7 @@ class TestAsyncCoordinatorBehavioral:
 
         sys.modules['python_multipart'].__version__ = '0.0.99'
         sys.modules['python_multipart.multipart'].parse_options_header = MagicMock(return_value={})
+        sys.modules['utils.log_sanitizer'].sanitize = lambda value: value
 
         _install_sync_observability_stubs()
 
@@ -1381,6 +1392,11 @@ class TestAsyncCoordinatorBehavioral:
         sys.modules['utils.byok'].get_byok_keys = MagicMock(return_value={})
         sys.modules['utils.analytics'].record_usage = MagicMock()
         sys.modules['utils.request_validation'].parse_sync_filename_timestamp = MagicMock(return_value=1700000000)
+        import types
+
+        sync_pkg = types.ModuleType('utils.sync')
+        sync_pkg.__path__ = [os.path.join(os.path.dirname(__file__), '..', '..', 'utils', 'sync')]
+        sys.modules['utils.sync'] = sync_pkg
         sys.modules['utils.sync'].files = sys.modules['utils.sync.files']
         sys.modules['utils.sync'].playback = sys.modules['utils.sync.playback']
         sys.modules['utils.sync.playback'].build_playback_artifact = MagicMock(return_value=b'')
@@ -1399,6 +1415,7 @@ class TestAsyncCoordinatorBehavioral:
         sys.modules['utils.subscription'].has_transcription_credits = MagicMock(return_value=True)
 
         sys.modules.pop('routers.sync', None)
+        sys.modules.pop('utils.sync.pipeline', None)
         import importlib.util
 
         spec = importlib.util.spec_from_file_location(
@@ -1408,10 +1425,13 @@ class TestAsyncCoordinatorBehavioral:
         module = importlib.util.module_from_spec(spec)
         spec.loader.exec_module(module)
 
+        pipeline_mod = sys.modules['utils.sync.pipeline']
+
         stubs['sync_jobs'] = mock_sync_jobs
         stubs['fair_use'] = sys.modules['utils.fair_use']
         stubs['byok'] = sys.modules['utils.byok']
         stubs['analytics'] = sys.modules['utils.analytics']
+        stubs['pipeline'] = pipeline_mod
         stubs['saved_modules'] = saved_modules
 
         return module, stubs
@@ -1419,6 +1439,7 @@ class TestAsyncCoordinatorBehavioral:
     @staticmethod
     def _cleanup(saved_modules):
         sys.modules.pop('routers.sync', None)
+        sys.modules.pop('utils.sync.pipeline', None)
         sys.modules.pop('sync_behavioral', None)
         for mod_name, orig in saved_modules.items():
             if orig is None:
@@ -1433,8 +1454,10 @@ class TestAsyncCoordinatorBehavioral:
 
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(side_effect=_HTTPException(status_code=400, detail='bad format'))
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(
+                side_effect=_HTTPException(status_code=400, detail='bad format')
+            )
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             await module._run_full_pipeline_background_async('j1', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job1')
 
@@ -1450,8 +1473,8 @@ class TestAsyncCoordinatorBehavioral:
         """Decode raising generic Exception must mark job failed."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(side_effect=RuntimeError('corrupt file'))
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(side_effect=RuntimeError('corrupt file'))
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             await module._run_full_pipeline_background_async('j2', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job2')
 
@@ -1466,8 +1489,8 @@ class TestAsyncCoordinatorBehavioral:
         """Empty wav_paths after decode must complete job with 0 segments."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=[])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=[])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             await module._run_full_pipeline_background_async('j3', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job3')
 
@@ -1483,20 +1506,20 @@ class TestAsyncCoordinatorBehavioral:
         """VAD errors must mark job failed and clean up."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _bad_vad(path, segmented_paths, errors):
                 errors.append(f'{path}: silero exploded')
 
-            module.retrieve_vad_segments = _bad_vad
+            stubs['pipeline'].retrieve_vad_segments = _bad_vad
 
             await module._run_full_pipeline_background_async('j4', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job4')
 
             stubs['sync_jobs'].mark_job_failed.assert_called()
             args = stubs['sync_jobs'].mark_job_failed.call_args[0]
             assert 'VAD failed' in args[1]
-            assert module._cleanup_files.call_count >= 2
+            assert stubs['pipeline']._cleanup_files.call_count >= 2
         finally:
             self._cleanup(stubs['saved_modules'])
 
@@ -1505,10 +1528,10 @@ class TestAsyncCoordinatorBehavioral:
         """Zero segmented_paths after VAD must complete with 0 segments."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
-            module.retrieve_vad_segments = MagicMock()
-            module.get_wav_duration = MagicMock(return_value=0.0)
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
+            stubs['pipeline'].retrieve_vad_segments = MagicMock()
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=0.0)
 
             await module._run_full_pipeline_background_async('j5', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job5')
 
@@ -1523,18 +1546,18 @@ class TestAsyncCoordinatorBehavioral:
         """DG budget exhausted must mark job failed."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _vad_with_segments(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_with_segments
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.FAIR_USE_ENABLED = True
-            module.FAIR_USE_RESTRICT_DAILY_DG_MS = 1000
-            module.get_enforcement_stage = MagicMock(return_value='restrict')
-            module.is_dg_budget_exhausted = MagicMock(return_value=True)
+            stubs['pipeline'].retrieve_vad_segments = _vad_with_segments
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].FAIR_USE_ENABLED = True
+            stubs['pipeline'].FAIR_USE_RESTRICT_DAILY_DG_MS = 1000
+            stubs['pipeline'].get_enforcement_stage = MagicMock(return_value='restrict')
+            stubs['pipeline'].is_dg_budget_exhausted = MagicMock(return_value=True)
             module.record_speech_ms = MagicMock()
             module.get_rolling_speech_ms = MagicMock(return_value={})
             module.check_soft_caps = MagicMock(return_value=[])
@@ -1552,32 +1575,32 @@ class TestAsyncCoordinatorBehavioral:
         """DG budget NOT exhausted must continue to segment processing."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _vad_with_segments(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_with_segments
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.FAIR_USE_ENABLED = True
-            module.FAIR_USE_RESTRICT_DAILY_DG_MS = 1000
-            module.get_enforcement_stage = MagicMock(return_value='restrict')
-            module.is_dg_budget_exhausted = MagicMock(return_value=False)
-            module.record_speech_ms = MagicMock()
-            module.get_rolling_speech_ms = MagicMock(return_value={})
-            module.check_soft_caps = MagicMock(return_value=[])
-            module.users_db = MagicMock()
-            module.users_db.get_user_transcription_preferences = MagicMock(return_value={})
-            module.build_person_embeddings_cache = MagicMock(return_value={})
-            module.process_segment = MagicMock()
-            module.record_dg_usage_ms = MagicMock()
-            module.record_usage = MagicMock()
+            stubs['pipeline'].retrieve_vad_segments = _vad_with_segments
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].FAIR_USE_ENABLED = True
+            stubs['pipeline'].FAIR_USE_RESTRICT_DAILY_DG_MS = 1000
+            stubs['pipeline'].get_enforcement_stage = MagicMock(return_value='restrict')
+            stubs['pipeline'].is_dg_budget_exhausted = MagicMock(return_value=False)
+            stubs['pipeline'].record_speech_ms = MagicMock()
+            stubs['pipeline'].get_rolling_speech_ms = MagicMock(return_value={})
+            stubs['pipeline'].check_soft_caps = MagicMock(return_value=[])
+            stubs['pipeline'].users_db = MagicMock()
+            stubs['pipeline'].users_db.get_user_transcription_preferences = MagicMock(return_value={})
+            stubs['pipeline'].build_person_embeddings_cache = MagicMock(return_value={})
+            stubs['pipeline'].process_segment = MagicMock()
+            stubs['pipeline'].record_dg_usage_ms = MagicMock()
+            stubs['pipeline'].record_usage = MagicMock()
 
             await module._run_full_pipeline_background_async('j7', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job7')
 
             stubs['sync_jobs'].mark_job_completed.assert_called_once()
-            module.process_segment.assert_called_once()
+            stubs['pipeline'].process_segment.assert_called_once()
         finally:
             self._cleanup(stubs['saved_modules'])
 
@@ -1586,23 +1609,25 @@ class TestAsyncCoordinatorBehavioral:
         """Partial segment failure must complete (not fail) with error count."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _vad_two_segments(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
                 segmented_paths.add('/tmp/seg_1700000002.wav')
 
-            module.retrieve_vad_segments = _vad_two_segments
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.users_db = MagicMock()
-            module.users_db.get_user_transcription_preferences = MagicMock(return_value={})
-            module.users_db.get_user_private_cloud_sync_enabled = MagicMock(return_value=False)
-            module.users_db.get_data_protection_level = MagicMock(return_value=None)
-            module.build_person_embeddings_cache = MagicMock(return_value={})
-            module._reprocess_merged_conversations = MagicMock()
-            module.record_usage = MagicMock()
-            module.get_timestamp_from_path = MagicMock(side_effect=lambda p: int(p.split('_')[-1].split('.')[0]))
+            stubs['pipeline'].retrieve_vad_segments = _vad_two_segments
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].users_db = MagicMock()
+            stubs['pipeline'].users_db.get_user_transcription_preferences = MagicMock(return_value={})
+            stubs['pipeline'].users_db.get_user_private_cloud_sync_enabled = MagicMock(return_value=False)
+            stubs['pipeline'].users_db.get_data_protection_level = MagicMock(return_value=None)
+            stubs['pipeline'].build_person_embeddings_cache = MagicMock(return_value={})
+            stubs['pipeline']._reprocess_merged_conversations = MagicMock()
+            stubs['pipeline'].record_usage = MagicMock()
+            stubs['pipeline'].get_timestamp_from_path = MagicMock(
+                side_effect=lambda p: int(p.split('_')[-1].split('.')[0])
+            )
             call_count = [0]
 
             def _process_seg_fails_once(path, uid, response, lock, errors, *args, **kwargs):
@@ -1612,7 +1637,7 @@ class TestAsyncCoordinatorBehavioral:
                 else:
                     response['new_memories'].add('mem1')
 
-            module.process_segment = _process_seg_fails_once
+            stubs['pipeline'].process_segment = _process_seg_fails_once
 
             await module._run_full_pipeline_background_async('j8', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job8')
 
@@ -1628,24 +1653,24 @@ class TestAsyncCoordinatorBehavioral:
         """A task-level segment exception must not complete with failed_segments=0."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _vad_one_segment(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_one_segment
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.users_db = MagicMock()
-            module.users_db.get_user_transcription_preferences = MagicMock(return_value={})
-            module.users_db.get_user_private_cloud_sync_enabled = MagicMock(return_value=False)
-            module.users_db.get_data_protection_level = MagicMock(return_value=None)
-            module.build_person_embeddings_cache = MagicMock(return_value={})
+            stubs['pipeline'].retrieve_vad_segments = _vad_one_segment
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].users_db = MagicMock()
+            stubs['pipeline'].users_db.get_user_transcription_preferences = MagicMock(return_value={})
+            stubs['pipeline'].users_db.get_user_private_cloud_sync_enabled = MagicMock(return_value=False)
+            stubs['pipeline'].users_db.get_data_protection_level = MagicMock(return_value=None)
+            stubs['pipeline'].build_person_embeddings_cache = MagicMock(return_value={})
             module._reprocess_merged_conversations = MagicMock()
-            module.record_usage = MagicMock()
+            stubs['pipeline'].record_usage = MagicMock()
             module.get_timestamp_from_path = MagicMock(return_value=1700000001)
             module.sanitize = lambda value: value
-            module.process_segment = MagicMock(side_effect=RuntimeError('executor exploded'))
+            stubs['pipeline'].process_segment = MagicMock(side_effect=RuntimeError('executor exploded'))
 
             await module._run_full_pipeline_background_async(
                 'j-generic-segment-error', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job-generic'
@@ -1664,25 +1689,25 @@ class TestAsyncCoordinatorBehavioral:
         """Person embeddings failure must fall back to empty dict, not crash."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _vad_one_seg(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_one_seg
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.users_db = MagicMock()
-            module.users_db.get_user_transcription_preferences = MagicMock(return_value={})
-            module.build_person_embeddings_cache = MagicMock(side_effect=RuntimeError('cache boom'))
+            stubs['pipeline'].retrieve_vad_segments = _vad_one_seg
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].users_db = MagicMock()
+            stubs['pipeline'].users_db.get_user_transcription_preferences = MagicMock(return_value={})
+            stubs['pipeline'].build_person_embeddings_cache = MagicMock(side_effect=RuntimeError('cache boom'))
             captured_cache = {}
 
             def _capture_process(path, uid, response, lock, errors, source, is_locked, prefs, cache, *args, **kwargs):
                 captured_cache['value'] = cache
                 response['new_memories'].add('m1')
 
-            module.process_segment = _capture_process
-            module.record_usage = MagicMock()
+            stubs['pipeline'].process_segment = _capture_process
+            stubs['pipeline'].record_usage = MagicMock()
 
             await module._run_full_pipeline_background_async('j9', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job9')
 
@@ -1696,18 +1721,18 @@ class TestAsyncCoordinatorBehavioral:
         """target_conversation_id must be forwarded to process_segment."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _vad_one_seg(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_one_seg
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.users_db = MagicMock()
-            module.users_db.get_user_transcription_preferences = MagicMock(return_value={})
-            module.build_person_embeddings_cache = MagicMock(return_value={})
-            module.record_usage = MagicMock()
+            stubs['pipeline'].retrieve_vad_segments = _vad_one_seg
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].users_db = MagicMock()
+            stubs['pipeline'].users_db.get_user_transcription_preferences = MagicMock(return_value={})
+            stubs['pipeline'].build_person_embeddings_cache = MagicMock(return_value={})
+            stubs['pipeline'].record_usage = MagicMock()
             captured_target = {}
 
             def _capture_target(
@@ -1716,7 +1741,7 @@ class TestAsyncCoordinatorBehavioral:
                 captured_target['value'] = target_cid
                 response['new_memories'].add('m1')
 
-            module.process_segment = _capture_target
+            stubs['pipeline'].process_segment = _capture_target
 
             await module._run_full_pipeline_background_async(
                 'j10', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job10', target_conversation_id='conv-123'
@@ -1731,26 +1756,26 @@ class TestAsyncCoordinatorBehavioral:
         """private_cloud_sync_enabled must be forwarded to process_segment."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
-            module._cleanup_files = MagicMock()
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline']._cleanup_files = MagicMock()
 
             def _vad_one_seg(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_one_seg
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.users_db = MagicMock()
-            module.users_db.get_user_transcription_preferences = MagicMock(return_value={})
-            module.users_db.get_user_private_cloud_sync_enabled = MagicMock(return_value=True)
-            module.build_person_embeddings_cache = MagicMock(return_value={})
-            module.record_usage = MagicMock()
+            stubs['pipeline'].retrieve_vad_segments = _vad_one_seg
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].users_db = MagicMock()
+            stubs['pipeline'].users_db.get_user_transcription_preferences = MagicMock(return_value={})
+            stubs['pipeline'].users_db.get_user_private_cloud_sync_enabled = MagicMock(return_value=True)
+            stubs['pipeline'].build_person_embeddings_cache = MagicMock(return_value={})
+            stubs['pipeline'].record_usage = MagicMock()
             captured = {}
 
             def _capture_private_sync(path, uid, response, lock, errors, *args, **kwargs):
                 captured['value'] = kwargs['private_cloud_sync_enabled']
                 response['new_memories'].add('m1')
 
-            module.process_segment = _capture_private_sync
+            stubs['pipeline'].process_segment = _capture_private_sync
 
             await module._run_full_pipeline_background_async(
                 'j-private', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job-private'
@@ -1765,20 +1790,20 @@ class TestAsyncCoordinatorBehavioral:
         """Cleanup must be called even on successful completion."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
             cleanup_calls = []
-            module._cleanup_files = lambda paths: cleanup_calls.append(list(paths))
+            stubs['pipeline']._cleanup_files = lambda paths: cleanup_calls.append(list(paths))
 
             def _vad_one_seg(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_one_seg
-            module.get_wav_duration = MagicMock(return_value=5.0)
-            module.users_db = MagicMock()
-            module.users_db.get_user_transcription_preferences = MagicMock(return_value={})
-            module.build_person_embeddings_cache = MagicMock(return_value={})
-            module.process_segment = MagicMock()
-            module.record_usage = MagicMock()
+            stubs['pipeline'].retrieve_vad_segments = _vad_one_seg
+            stubs['pipeline'].get_wav_duration = MagicMock(return_value=5.0)
+            stubs['pipeline'].users_db = MagicMock()
+            stubs['pipeline'].users_db.get_user_transcription_preferences = MagicMock(return_value={})
+            stubs['pipeline'].build_person_embeddings_cache = MagicMock(return_value={})
+            stubs['pipeline'].process_segment = MagicMock()
+            stubs['pipeline'].record_usage = MagicMock()
 
             await module._run_full_pipeline_background_async('j11', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job11')
 
@@ -1792,15 +1817,15 @@ class TestAsyncCoordinatorBehavioral:
         """Cleanup and BYOK clear must happen even when pipeline crashes."""
         module, stubs = self._load_sync_module()
         try:
-            module.decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
+            stubs['pipeline'].decode_files_to_wav = MagicMock(return_value=['/tmp/w.wav'])
             cleanup_calls = []
-            module._cleanup_files = lambda paths: cleanup_calls.append(list(paths))
+            stubs['pipeline']._cleanup_files = lambda paths: cleanup_calls.append(list(paths))
 
             def _vad_one_seg(path, segmented_paths, errors):
                 segmented_paths.add('/tmp/seg_1700000001.wav')
 
-            module.retrieve_vad_segments = _vad_one_seg
-            module.get_wav_duration = MagicMock(side_effect=RuntimeError('unexpected crash'))
+            stubs['pipeline'].retrieve_vad_segments = _vad_one_seg
+            stubs['pipeline'].get_wav_duration = MagicMock(side_effect=RuntimeError('unexpected crash'))
 
             await module._run_full_pipeline_background_async('j12', 'uid', ['/tmp/f.opus'], 'omi', False, '/tmp/job12')
 
@@ -1867,7 +1892,6 @@ class TestV2EndpointExecution:
             'utils.log_sanitizer',
             'utils.http_client',
             'utils.request_validation',
-            'utils.sync',
             'utils.sync.files',
             'utils.sync.playback',
             'utils.speaker_assignment',
@@ -1883,6 +1907,7 @@ class TestV2EndpointExecution:
 
         sys.modules['python_multipart'].__version__ = '0.0.99'
         sys.modules['python_multipart.multipart'].parse_options_header = MagicMock(return_value={})
+        sys.modules['utils.log_sanitizer'].sanitize = lambda value: value
 
         _install_sync_observability_stubs()
 
@@ -1915,6 +1940,11 @@ class TestV2EndpointExecution:
         sys.modules['utils.fair_use'].FAIR_USE_RESTRICT_DAILY_DG_MS = 0
         sys.modules['utils.subscription'].has_transcription_credits = MagicMock(return_value=True)
         sys.modules['utils.request_validation'].parse_sync_filename_timestamp = MagicMock(return_value=1700000000)
+        import types
+
+        sync_pkg = types.ModuleType('utils.sync')
+        sync_pkg.__path__ = [os.path.join(os.path.dirname(__file__), '..', '..', 'utils', 'sync')]
+        sys.modules['utils.sync'] = sync_pkg
         sys.modules['utils.sync'].files = sys.modules['utils.sync.files']
         sys.modules['utils.sync'].playback = sys.modules['utils.sync.playback']
         sys.modules['utils.sync.playback'].build_playback_artifact = MagicMock(return_value=b'')
@@ -1937,6 +1967,7 @@ class TestV2EndpointExecution:
     @staticmethod
     def _cleanup_modules(saved_modules):
         sys.modules.pop('routers.sync', None)
+        sys.modules.pop('utils.sync.pipeline', None)
         for mod_name, orig in saved_modules.items():
             if orig is None:
                 sys.modules.pop(mod_name, None)
@@ -1950,6 +1981,7 @@ class TestV2EndpointExecution:
 
         try:
             sys.modules.pop('routers.sync', None)
+            sys.modules.pop('utils.sync.pipeline', None)
             import importlib.util
 
             spec = importlib.util.spec_from_file_location(
@@ -1992,6 +2024,7 @@ class TestV2EndpointExecution:
 
         try:
             sys.modules.pop('routers.sync', None)
+            sys.modules.pop('utils.sync.pipeline', None)
             import importlib.util
 
             spec = importlib.util.spec_from_file_location(
@@ -2033,6 +2066,7 @@ class TestV2EndpointExecution:
 
         try:
             sys.modules.pop('routers.sync', None)
+            sys.modules.pop('utils.sync.pipeline', None)
             import importlib.util
 
             spec = importlib.util.spec_from_file_location(
@@ -2079,6 +2113,7 @@ class TestV2EndpointExecution:
 
         try:
             sys.modules.pop('routers.sync', None)
+            sys.modules.pop('utils.sync.pipeline', None)
             import importlib.util
 
             spec = importlib.util.spec_from_file_location(
@@ -2124,6 +2159,7 @@ class TestV2EndpointExecution:
 
         try:
             sys.modules.pop('routers.sync', None)
+            sys.modules.pop('utils.sync.pipeline', None)
             import importlib.util
 
             spec = importlib.util.spec_from_file_location(
@@ -2166,6 +2202,7 @@ class TestV2EndpointExecution:
 
         try:
             sys.modules.pop('routers.sync', None)
+            sys.modules.pop('utils.sync.pipeline', None)
             import importlib.util
 
             spec = importlib.util.spec_from_file_location(
@@ -2368,7 +2405,7 @@ class TestBYOKContextPropagation:
 
     def test_async_coordinator_clears_byok_in_finally(self):
         """Async coordinator must clear BYOK context in its finally block."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
         next_boundary = source.find('\n@router.', start + 1)
         if next_boundary == -1:
@@ -2380,7 +2417,7 @@ class TestBYOKContextPropagation:
 
     def test_async_coordinator_sets_byok_uid_before_work(self):
         """Async coordinator must attach the uid to inherited BYOK key context."""
-        source = self._read_sync_source()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
         next_boundary = source.find('\n@router.', start + 1)
         if next_boundary == -1:
@@ -2513,11 +2550,11 @@ class TestTimeoutConfiguration:
 
     def test_segment_timeout_budget(self):
         """Segment tasks in v2 async coordinator must use asyncio.wait_for with timeout=300."""
-        sync_path = os.path.join(os.path.dirname(__file__), '..', '..', 'routers', 'sync.py')
-        with open(sync_path, encoding='utf-8') as f:
-            source = f.read()
+        source = _read_pipeline_source()
         start = source.index('async def _run_full_pipeline_background_async')
-        end = source.find('\n@router.', start + 1)
+        end = source.find('\nasync def ', start + 1)
+        if end == -1:
+            end = source.find('\ndef ', start + 1)
         if end == -1:
             end = len(source)
         func_body = source[start:end]
